Use the absolute path for compiler arguments
We use make for our build system.  Our make is somewhat primitive
and doesn't have rules for indiviual files.  Thus, bear is used to
intercept these calls so a compile_commands.json file could be
produced which allows the use of building individual targets.  Bear
uses relative paths to the single file when creating these files
but this doesn't play well with tools like clang-tidy.  This is
because clang-tidy will search recursively for a .clang-tidy file
but only up to the final directory of the file.  Without an absolute
path, the search ends at the relative path of the file.

Signed-off-by: Christian Venegas <cvenegas@esri.com>